chore: lazy load playground
