+++ playground/src/lib/config/eval-config.ts
@@ -18,7 +18,6 @@ export const evalConfig = (config: string, _scope?: Dict): PlaygroundConfig | nu
   try {
     const scope = Object.assign({}, pandaDefs, _scope)
     const config = evalCode(`return (() => {${codeTrimmed}; return config})()`, scope)
-    if (!_scope) delete config.presets
     return config
   } catch (e) {
     return null
